tensorflow except for hessians
