+++ pennylane/interfaces/tensorflow_autograph.py
@@ -26,10 +26,9 @@ from pennylane.measurements import SampleMP, StateMP
 
 from .tensorflow import (
     _compute_vjp,
-    _jac_restructured,
     _res_restructured,
     _to_tensors,
-    set_parameters_on_copy_and_unwrap,
+    set_parameters_on_copy,
 )
 
 
@@ -41,6 +40,30 @@ def _flatten_nested_list(x):
         return [x]
 
     return reduce(lambda a, y: a + _flatten_nested_list(y), x, [])
+
+
+def _jac_restructured(jacs, tapes):
+    """
+    Reconstruct the nested tuple structure of the jacobian of a list of tapes
+    """
+    start = 0
+    jacs_nested = []
+    for tape in tapes:
+        num_meas = len(tape.measurements)
+        num_params = len(tape.trainable_params)
+
+        tape_jacs = tuple(jacs[start : start + num_meas * num_params])
+        tape_jacs = tuple(
+            tuple(tape_jacs[i * num_params : (i + 1) * num_params]) for i in range(num_meas)
+        )
+
+        while isinstance(tape_jacs, tuple) and len(tape_jacs) == 1:
+            tape_jacs = tape_jacs[0]
+
+        jacs_nested.append(tape_jacs)
+        start += num_meas * num_params
+
+    return tuple(jacs_nested)
 
 
 def execute(
@@ -132,7 +155,7 @@ def execute(
         params_unwrapped = _nest_params(all_params)
         output_sizes = []
 
-        new_tapes = set_parameters_on_copy_and_unwrap(tapes, params_unwrapped, unwrap=False)
+        new_tapes = set_parameters_on_copy(tapes, params_unwrapped)
         # Forward pass: execute the tapes
         res, jacs = execute_fn(new_tapes, **gradient_kwargs)
 
@@ -211,9 +234,7 @@ def execute(
 
                             dy = _res_restructured(dy, tapes)
 
-                            new_tapes = set_parameters_on_copy_and_unwrap(
-                                tapes, params_unwrapped, unwrap=False
-                            )
+                            new_tapes = set_parameters_on_copy(tapes, params_unwrapped)
                             vjp_tapes, processing_fn = qml.gradients.batch_vjp(
                                 new_tapes,
                                 dy,
@@ -278,9 +299,7 @@ def execute(
                         all_params = all_params[:len_all_params]
                         params_unwrapped = _nest_params(all_params)
 
-                        new_tapes = set_parameters_on_copy_and_unwrap(
-                            tapes, params_unwrapped, unwrap=False
-                        )
+                        new_tapes = set_parameters_on_copy(tapes, params_unwrapped)
                         jac = gradient_fn(new_tapes, **gradient_kwargs)
 
                         vjps = _compute_vjp(dy, jac, multi_measurements, has_partitioned_shots)
